Create a system of equation solver
Since circuits have a unique solution, all of its equations can be represented as A*x = b.
Where A is an NxN matrix of the variables' coefficients, b holds the constants, and x is the vector of variables we are solving for.
If we solve for R=rref(A) and make the corresponding updates to b, then we would have an equivalent equation R*x = d where d is the corresponding RHS of the equations.
Since A is NxN, R=I, therefore we can read off the solution x = d.

Algorithm outline for finding rref(A) was referenced from https://www.math.purdue.edu/~shao92/documents/Algorithm%20REF.pdf